Add workaround for windbg trying to read windows shared data page (#21)
## Description

Windbg tries to read the Windows shared data page as some sort of
heuristic when connecting. Because this page does not exist in UEFI
mappings, Windows may loop on this read so long as it fails. This work
around will just return 0 when reads occur to this well-known VA so
cause the heuristic to fail fast.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested on Q35 virtual platform.

## Integration Instructions

N/A