Implement hardware watchpoints or data breakpoints (#37)
## Description

Implement up to 4 watchpoints, or data breakpoints, which are
represented by the GDB breakpoints types 2-4. These can be set in windbg
by using the `ba` command, e.g. `ba w8 0x10000001`. These can be set in
GDB by using the `watch` for writes, `rwatch` for reads, or `awatch` for
either.

#33 


- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested on Q35 and SBSA on QEMU.

## Integration Instructions

N/A