UefiExt: Add Option To Ignore Self Map in !pt
Currently !pt will always search for a self map entry
and use that to walk the page tables. However, this doesn't
let the extension walk uninstalled page tables (because that
virtual address range is not mapped).

This adds a flag to !pt to ignore the self map and instead parse
the page tables in a standard walk to handle the offline page
table parsing case.